edición de proyectos, consulta de uno por id y eliminación de proyectos

+++ src/app/api/projects/[id]/route.ts
@@ -70,18 +70,39 @@ export async function PUT(
 
 export async function DELETE(
 	request: NextRequest,
-	{ params }: { params: Promise<{ id: string }> }) {
-	
+	{ params }: { params: Promise<{ id: string }> },
+) {
 	try {
-		const {id} = await params;
-		
-	  await prisma.projects.delete({
-		where: { id: Number(id) },
-	  });
-  
-	  return NextResponse.json({ message: "Proyecto eliminado correctamente" });
+		const { id } = await params;
+		if (!id) {
+			return NextResponse.json({ error: 'ID inválido' }, { status: 400 });
+		}
+
+		// Verificar si el proyecto existe antes de eliminarlo
+		const project = await prisma.projects.findUnique({
+			where: { id: +id },
+		});
+
+		if (!project) {
+			return NextResponse.json(
+				{ error: 'Proyecto no encontrado' },
+				{ status: 404 },
+			);
+		}
+
+		// Eliminar el proyecto
+		await prisma.projects.delete({
+			where: { id: +id },
+		});
+
+		return NextResponse.json({
+			message: 'Proyecto eliminado correctamente',
+		});
 	} catch (error) {
-		console.log(error);
-	  return NextResponse.json({ error: "Error al eliminar el proyecto" }, { status: 500 });
+		console.error('Error al eliminar el proyecto:', error);
+		return NextResponse.json(
+			{ error: 'Error al eliminar el proyecto' },
+			{ status: 500 },
+		);
 	}
-  }
+}
